Move multisigAccountDetails
