+++ src/modules/wallet/components/multisigAccountDetails/index.test.js
