+++ src/modules/wallet/components/multisigAccountDetails/styles.css
